refactor(cli): move the --override-count functions out of cmd/helpers.go
ApplyCountOverride, evidencedCount and reportCountOverride move verbatim into
cmd/helpers_count_override.go. The public signature is unchanged and every
caller is in the same package, so nothing else moves and nothing is renamed.

This is a pure move, verified as one rather than asserted: every added line
except the six-line package and import header is byte-identical to a removed
line, in the same order, and the source shrank by exactly the moved block plus
its one blank separator. No logic, signature, name or comment changed. The
three functions reference only common.Recommendation, common.IsSavingsPlan,
common.ScaleRecommendationCosts and the package-level AppLogger, so the new
file needs a single import and the source loses none.

cmd/helpers.go goes from 881 to 794 lines, which is 19 below the 813 it stood
at before this PR touched it: the 68 lines this branch added are back out
along with the function they replaced. The new file is 92 lines.

The file is still over the 500-line limit, as are five other non-test files in
cmd/. That is pre-existing and systemic, tracked in #1834, and restructuring
813 lines of money-path code inside a correctness fix is the wrong risk and
makes the behavior diff unreadable. Taking this branch's own additions back
out is the part that belongs here.

Test parity is asserted on the SET of test names and outcomes rather than on a
count, comparing a run in a pristine worktree at the pre-move commit against
one on this tree: identical names, identical outcomes, no additions, no
removals. golangci-lint at the CI-pinned v2.10.1 reports 0 issues and gocyclo
-over 10 is clean on both files, checked after the move because moving code
changes which file the complexity is attributed to.
